requirements.txt
fastapi
uvicorn
python-telegram-bot==20.*
google-api-python-client
google-auth
fastapi
uvicorn[standard]
python-dotenv
httpx
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib